Rename all classes of workers
