+++ src/main/java/core/ReactiveWorker_V3F.java
@@ -3,24 +3,31 @@ package core;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 
-/** Worker that must defines two task (method) that run at the same time,
- * and whose results of the first feed the second. The second task return the result
- * to launch method that return to caller.
- * @param <T> : is the type of exchanged data between first task and second task */
-public abstract class TripleComputer<T,U,V> implements Runnable {
+/** Worker that must defines three tasks (method).
+ * The results of this tasks are collected in a fifo.
+ * This fifo can be read by launcher (caller) using next() method.
+ * First and Second task work in same time, and results of first task are sended to second task through
+ * an internal fifo.
+ * Second and Third task work in same time too, and results of second task are sended to third task through
+ * a second internal fifo.
+ * @param <T> : is the type exchanged between first and second task
+ * @param <U> : is the type exchanged between second and third task
+ * @param <V> : is the type returned to the caller by second task */
+public abstract class ReactiveWorker_V3F<T,U,V> implements Runnable {
 
     private int taskToLaunch = 1;               // Counter to launch each task only once.
     private BlockingQueue<T> internalFifo1;     // The internal fifo between first task  and second task
-    private BlockingQueue<U> internalFifo2;     // The output fifo into the second task write, and caller read
+    private BlockingQueue<U> internalFifo2;     // The internal fifo between second task and third task
+    private BlockingQueue<V> outPutFifo;        // The output fifo into the second task write, and caller read
     private boolean firstHasFinished  = false;  // true when the first task has finished
     private boolean secondHasFinished = false;  // true when the second task has finished
     private boolean thirdHasFinished  = false;  // true when the second task has finished
-    private V computedResult = null;
 
-    /** Starts the first and second task in their own thread */
-    public final V launch(){
+    /** Starts the first, second and third task in their own thread */
+    public final void launch(){
         internalFifo1 = new ArrayBlockingQueue<T>(10000);
         internalFifo2 = new ArrayBlockingQueue<U>(10000);
+        outPutFifo = new ArrayBlockingQueue<V>(10000);
 
         Thread t1 = new Thread(this);
         t1.start();
@@ -28,18 +35,8 @@ public abstract class TripleComputer<T,U,V> implements Runnable {
         t2.start();
         Thread t3 = new Thread(this);
         t3.start();
-        while( thirdHasFinished == false){
-            try {
-                Thread.sleep(1);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-
-        return computedResult;
     }
 
-    /** Launch tasks and indicates when each is completed */
     @Override
     public final void run() {
         int numberTaskToLaunch = 0;
@@ -70,7 +67,7 @@ public abstract class TripleComputer<T,U,V> implements Runnable {
     abstract void firstTask();
     /** Second task to execute. This second task has take results of first task using next() method, and used them */
     abstract void secondTask();
-    /** Second task to execute. This second task has take results of first task using next() method, and used them */
+    /** Third task to execute. This third task has take results of second task using next() method, and used them */
     abstract void thirdTask();
 
     /** send the specified element to the secondTask.
@@ -79,7 +76,7 @@ public abstract class TripleComputer<T,U,V> implements Runnable {
     protected void toSecondTask(T element) throws InterruptedException {
         internalFifo1.put(element);
     }
-    /** send the specified element to the secondTask.
+    /** send the specified element to the thirdTask.
      * Some exceptions can be throwed, if element is null, if element contains some attributs that prevent it
      * to be putted into the queue, etc... */
     protected void toThirdTask(U element) throws InterruptedException {
@@ -94,6 +91,7 @@ public abstract class TripleComputer<T,U,V> implements Runnable {
             if (firstHasFinished) return null;
         }
     }
+
     /** @return The next element from the fifo. If fifo is empty and producer is dead, then null is returned */
     protected U fromSecondTask(){
         while(true) {
@@ -103,8 +101,21 @@ public abstract class TripleComputer<T,U,V> implements Runnable {
         }
     }
 
-    /** return the given result to the caller of this object */
-   protected void returnResult(V result){
-        computedResult = result;
-   }
+    /** send the specified element through main fifo (to the launcher of this object).
+     * Some exceptions can be throwed, if element is null, if element contains some attributs that prevent it
+     * to be putted into the queue, etc... */
+    protected void answer(V element) throws InterruptedException {
+        outPutFifo.put(element);
+    }
+
+    /** This method has to be used by the caller (which has called launch() methode).
+     * @return The next element from the fifo filled by the second task */
+    public V next(){
+        while(true) {
+            V element = outPutFifo.poll();
+            if (null != element) return element;
+            if (thirdHasFinished) return null;
+        }
+    }
+
 }
